+++ packages/modal/src/index.tsx
@@ -1,27 +1,61 @@
 import * as React from 'react';
 import Overlay from '@lendi-ui/overlay';
 import { Fade } from '@lendi-ui/transition';
-import { Wrapper, Container, Content, Footer, ModalSize, CloseIcon } from './index.style';
+import {
+  HeadingSize,
+  Wrapper,
+  Container,
+  HeaderWrapper,
+  TitlesWrapper,
+  Title,
+  Subtitle,
+  Content,
+  Footer,
+  ModalSize,
+  CloseIcon,
+} from './index.style';
 import { LUIGlobalProps } from '@lendi-ui/utils';
 
 export interface ModalProps extends LUIGlobalProps {
   size?: ModalSize;
+  fixedHeader?: boolean;
   show: boolean;
   onHide: () => void;
   children?: React.ReactNode;
 }
 
+export interface ModalHeaderProps {
+  title: string;
+  headerSize?: HeadingSize;
+  subtitle?: string;
+}
+
+export const Header = (props: ModalHeaderProps) => {
+  const size = props.headerSize || 'sm';
+
+  return (
+    <HeaderWrapper>
+      <TitlesWrapper>
+        {props.title && <Title size={size}>{props.title}</Title>}
+        {props.subtitle && <Subtitle>{props.subtitle}</Subtitle>}
+      </TitlesWrapper>
+    </HeaderWrapper>
+  );
+};
+
 class Modal extends React.Component<ModalProps> {
+  public static Header = Header;
   public static Content = Content;
   public static Footer = Footer;
 
   render() {
-    const { show, size, onHide, children, ...otherProps } = this.props;
+    const { show, size, onHide, fixedHeader = false, children, ...otherProps } = this.props;
+
     return (
       <Wrapper>
         <Overlay show={show} zIndex={5} onHide={onHide} />
         <Fade active={show}>
-          <Container show={show} size={size} className="modal-container" {...otherProps}>
+          <Container show={show} size={size} fixedHeader={fixedHeader} className="modal-container" {...otherProps}>
             <CloseIcon color="shade.300" onClick={onHide} />
             {children}
           </Container>